+++ src/assets/colors.ts
@@ -4,4 +4,26 @@ export const COLORS = {
 
     mainGradientTop: '#dddddd',
     mainGradientBottom: '#7e7e7e',
+
+
+}
+
+export const COLORTHEME = {
+    yellow: {
+        primary: '#FFFF99',
+        seacondary: '#FFCC00',
+        tertiary: '#FF9900'
+    },
+
+    red: {
+        primary: '#FF9999',
+        seacondary: '#FF6666',
+        tertiary: '#FF3333'
+    },
+
+    blue: {
+        primary: '#99CCFF',
+        seacondary: '#6699FF',
+        tertiary: '#3366FF'
+    }
 }