+++ src/interfaces/api.interface.ts
@@ -1,0 +1,4 @@
+export interface ITemperature {
+    celsius: number,
+    fahrenheit: number
+}