+++ src/interfaces/temperatureView.interface.ts
@@ -1,0 +1,8 @@
+import { ITemperature } from "./api.interface";
+import { TemperaturaScale } from "./types";
+
+export interface ITemperatureViewProps {
+    temperature: ITemperature;
+    day: string;
+    temperatureScale: TemperaturaScale;
+}