+++ src/interfaces/text.interface.ts
@@ -1,0 +1,7 @@
+import { IChildrenProps } from "./childrenProps.interface"
+import { TextType } from "./types";
+
+export interface ITextProps extends IChildrenProps {
+    className?: string;
+    as?: TextType;
+}