+++ src/interfaces/types.ts
@@ -1,0 +1,8 @@
+export type ColorTheme = "red" | "blue" | "yellow"
+
+export type TextType = "h1" | "h2" | "h3"
+
+export type TemperaturaScale = "celsius" | "fahrenheit"
+
+export type WindDirection = "N" | "L" | "S" | "O" | "NE" | "SE" | "SO" | "NO"
+